Add files via upload
This was an exciting but difficult task for me as I was extracting an hyperlink from a message. I had to duplicate the column that contained the message and the hyperlink to keep the original message intact. then, split column with the delimiter "http://", write conditions to it, rename columns, merge columns, trim the columns to remove any whitespace. Kindly take a look at it and let me know what you think. Give me a feedback to build myself. Thank you.